Modify stopping conditions for adding enemies
Modify the stopping conditions for how enemies are added when the number is not specified.  Previously, we would essentially max out the XP for a given category, which isn't realistic.  Now, unless a specific number of enemies was requested, once we've met the minimum requirements, there is a 50/50 chance we'll stop.
Also continued development of run_encounters.py